+++ src/stirling/source_connectors/socket_tracer/BUILD.bazel
@@ -183,7 +183,7 @@ pl_cc_bpf_test(
         "//src/common/testing/test_utils:cc_library",
         "//src/stirling/source_connectors/socket_tracer/testing:cc_library",
         "//src/stirling/source_connectors/socket_tracer/testing/container_images:curl_container",
-        "//src/stirling/source_connectors/socket_tracer/testing/container_images:nginx_open_ssl_1_1_1_container",
+        "//src/stirling/source_connectors/socket_tracer/testing/container_images:nginx_openssl_1_1_1_container",
         "//src/stirling/testing:cc_library",
     ],
 )
@@ -454,9 +454,9 @@ pl_cc_bpf_test(
         "//src/common/testing/test_utils:cc_library",
         "//src/stirling/source_connectors/socket_tracer/testing:cc_library",
         "//src/stirling/source_connectors/socket_tracer/testing/container_images:curl_container",
-        "//src/stirling/source_connectors/socket_tracer/testing/container_images:nginx_open_ssl_1_1_0_container",
-        "//src/stirling/source_connectors/socket_tracer/testing/container_images:nginx_open_ssl_1_1_1_container",
-        "//src/stirling/source_connectors/socket_tracer/testing/container_images:nginx_open_ssl_3_0_7_container",
+        "//src/stirling/source_connectors/socket_tracer/testing/container_images:nginx_openssl_1_1_0_container",
+        "//src/stirling/source_connectors/socket_tracer/testing/container_images:nginx_openssl_1_1_1_container",
+        "//src/stirling/source_connectors/socket_tracer/testing/container_images:nginx_openssl_3_0_8_container",
         "//src/stirling/source_connectors/socket_tracer/testing/container_images:node_12_3_1_container",
         "//src/stirling/source_connectors/socket_tracer/testing/container_images:node_14_18_1_alpine_container",
         "//src/stirling/source_connectors/socket_tracer/testing/container_images:node_client_container",
@@ -499,7 +499,7 @@ pl_cc_bpf_test(
         ":cc_library",
         "//src/common/testing/test_utils:cc_library",
         "//src/stirling/source_connectors/socket_tracer/testing:cc_library",
-        "//src/stirling/source_connectors/socket_tracer/testing/container_images:nginx_open_ssl_1_1_0_container",
+        "//src/stirling/source_connectors/socket_tracer/testing/container_images:nginx_openssl_1_1_0_container",
         "//src/stirling/source_connectors/socket_tracer/testing/container_images:ruby_container",
         "//src/stirling/testing:cc_library",
     ],
