+++ src/stirling/source_connectors/socket_tracer/dyn_lib_trace_bpf_test.cc
@@ -28,7 +28,7 @@
 #include "src/shared/types/column_wrapper.h"
 #include "src/shared/types/types.h"
 #include "src/stirling/source_connectors/socket_tracer/socket_trace_connector.h"
-#include "src/stirling/source_connectors/socket_tracer/testing/container_images/nginx_open_ssl_1_1_0_container.h"
+#include "src/stirling/source_connectors/socket_tracer/testing/container_images/nginx_openssl_1_1_0_container.h"
 #include "src/stirling/source_connectors/socket_tracer/testing/container_images/ruby_container.h"
 #include "src/stirling/source_connectors/socket_tracer/testing/protocol_checkers.h"
 #include "src/stirling/source_connectors/socket_tracer/testing/socket_trace_bpf_test_fixture.h"
