+++ src/stirling/source_connectors/socket_tracer/testing/container_images/nginx_openssl_1_1_0_container.h
