+++ src/stirling/source_connectors/socket_tracer/testing/container_images/nginx_openssl_1_1_1_container.h
