+++ src/stirling/source_connectors/socket_tracer/testing/containers/BUILD.bazel
@@ -84,8 +84,8 @@ container_image(
 )
 
 container_image(
-    name = "nginx_alpine_openssl_3_0_7_image",
-    base = "@nginx_alpine_openssl_3_0_7_base_image//image",
+    name = "nginx_alpine_OpenSSL_3_0_8_image",
+    base = "@nginx_alpine_OpenSSL_3_0_8_base_image//image",
     layers = [
         "//src/stirling/source_connectors/socket_tracer/testing/containers/ssl:nginx_conf",
         "//src/stirling/source_connectors/socket_tracer/testing/containers/ssl:nginx_html",
